Release 0.54a
Reorganized project structure, made the following changes:

Added orbital waypoints that show a ring and are animated to show you
which direction the orbit goes. These show up in the tracking station
for all currently posted contracts, and in game on the map for any that
you've accepted, just like normal waypoints.

Added satellite deployment contracts. They do not show up until you have
technology for probe cores, power sources, and antennae. These contracts
will ask you to build a new unmanned probe with an antenna and a power
source, then put it into a very specific orbit, shown on the map with
the new waypoints. The mission will ask you to put it into a certain
deviation of the orbit shown, and the deviation gets smaller as the
difficulty goes up. You need to match everything: apoapsis, periapsis,
inclination, eccentricity, the whole nine yards. The orbits also get
less kosher as the difficulty goes up, more eccentric and inclined, and
the burns required to hit that deviation window are very minute. It's
challenging, but the pay goes up as well. This mission type prefers
Kerbin, but can go to other planets as well, especially at higher
difficulties. On Kerbin, the reputation and fund rewards are tripled,
but the science is cut to a third. This is to emulate a commercial space
program on Kerbin putting up satellites for clients. That doesn't mean
it will pay as well as putting a satellite around Moho, it just means it
will be comparable.

Added new "neutralize controls for a period of time" objective. This
objective requires you to cut your throttle, and turn off any controls,
such as RCS or SAS for a certain period of time. This was added to
asteroid recovery missions, base building missions, station building
missions, and satellite deployment missions to prevent the mission from
completing before the situation has...stabilized.

Adjusted the way that distance calculations to waypoints are calculated,
specifically in regards to altitude. For aerial waypoints this means
that you can now fly anywhere within the height envelope shown on the
contract and still complete them. For rover contracts, it means that if
a waypoint spawns at an odd height, it doesn't matter, as you just need
to drive to the general area.

Drastically reduced the distance required to trigger rover waypoints.
This was made possible by the previous change, and means that you now
have to actually drive around within a rover cluster.

Fixed the cupola objective plaguing some station missions, a big thanks
to Burkitt for helping me track this issue down.