Release 0.57b
ARM contracts doomed for Kerbin will now occasionally ask you to land
instead of orbit with the asteroid.

Stationary orbits now ask for you to orbit over a specific point with
line of sight to it. Be within 45 degrees of that point to accomplish
this objective. It will be shown as a new waypoint.

To compensate for the added difficulty of stationary orbits, they were
raised to exceptional prestige level, which has the added benefits of
making them worth more, and making synchronous orbits more plentiful in
the lower difficulties.

Argument of periapsis checks are now skipped for perfectly round orbits.

Ascending and descending nodes now show the proper relative inclination
differences.

There is currently a known issue with the board reset upon patching.
Patching may or may not reset your contracts. If it does, any active
contracts will be refreshed on the board upon patching. This is
intended. If it does not, you will have goofy contracts for a few days
of in game time. This is not intended. If you see any satellite
contracts with 400KM apoapsis and periapsis, decline them so that new
ones can generate. You should only see them once, if at all.